pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Gabriel Martin del Campo", email = "gmartindelcye@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

